% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{compter_nombre_d_adjoints}
\alias{compter_nombre_d_adjoints}
\title{Compter le nombre d'adjoints}
\usage{
compter_nombre_d_adjoints(df)
}
\arguments{
\item{df}{Un dataframe qui valide le schéma contenant les données
des élus conseillers municipaux.}
}
\value{
Un entier représentant le nombre total d'adjoints dans le dataframe
}
\description{
Cette fonction calcule le nombre total d'adjoints au maire à partir d'un
dataframe respectant le schéma des données des élus conseillers municipaux.
Le dataframe doit contenir les mêmes colonnes que le fichier
élus-conseillers-municipaux-cm.
}
