% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Données des élus conseillers municipaux en France}
\format{
\subsection{\code{dataset}}{

Un data frame avec 489973 lignes et 16 colonnes :
\describe{
\item{Code.du.département}{Code INSEE du département.}
\item{Libellé.du.département}{Nom du département.}
\item{Code.de.la.collectivité.à.statut.particulier}{Code de la collectivité avec un statut particulier.}
\item{Libellé.de.la.collectivité.à.statut.particulier}{Nom de la collectivité avec un statut particulier.}
\item{Code.de.la.commune}{Code INSEE de la commune.}
\item{Libellé.de.la.commune}{Nom de la commune.}
\item{Nom.de.l.élu}{Nom de l'élu municipal.}
\item{Prénom.de.l.élu}{Prénom de l'élu municipal.}
\item{Code.sexe}{Code du sexe de l'élu (M/F).}
\item{Date.de.naissance}{Date de naissance de l'élu.}
\item{Code.de.la.catégorie.socio.professionnelle}{Code de la catégorie socio-professionnelle.}
\item{Libellé.de.la.catégorie.socio.professionnelle}{Libellé de la catégorie socio-professionnelle.}
\item{Date.de.début.du.mandat}{Date de début du mandat de l'élu.}
\item{Libellé.de.la.fonction}{Fonction occupée par l'élu.}
\item{Date.de.début.de.la.fonction}{Date de début de la fonction.}
\item{Code.nationalité}{Code de la nationalité de l'élu.}
}
}
}
\source{
Répertoire National des Élus, data.gouv.fr, consulté en février 2025 : \url{https://www.data.gouv.fr/fr/datasets/repertoire-national-des-elus-1/#/resources/d5f400de-ae3f-4966-8cb6-a85c70c6c24a}
}
\usage{
dataset
}
\description{
Ce jeu de données contient des informations sur les élus conseillers municipaux
en France, incluant les détails sur les départements, communes, élus, et mandats.
}
\keyword{datasets}
