% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport.R
\name{generer_rapport}
\alias{generer_rapport}
\title{Générer un rapport à partir du fichier rapport.qmd}
\usage{
generer_rapport(commune, departement, output)
}
\arguments{
\item{commune}{Code INSEE de la commune pour laquelle générer le rapport.}

\item{departement}{Code INSEE du département pour lequel générer le rapport.}

\item{output}{Chemin du fichier de sortie (par exemple, "rapport_final.html").}
}
\value{
Aucune valeur de retour, mais génère un fichier HTML à l'emplacement spécifié par \code{output}.
}
\description{
Cette fonction génère un rapport HTML basé sur les paramètres fournis pour une commune et un département.
Elle utilise Quarto pour compiler le rapport.
}
\details{
La fonction utilise le fichier \code{rapport.qmd} situé dans le dossier \code{inst} du package.
Les paramètres \code{commune} et \code{departement} sont passés au rapport pour configurer les données à afficher.
}
\examples{
\dontrun{
generer_rapport(commune = "44109", departement = "44", output = "rapport_final.html")
}

}
