% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{plot.commune}
\alias{plot.commune}
\title{Graphique du nombre d'élus par catégorie socio professionnelle par commune}
\usage{
\method{plot}{commune}(df)
}
\arguments{
\item{df}{Un dataframe qui contient une seule commune et qui valide le schéma
contenant les données des élus conseillers municipaux.}
}
\value{
La fonction retourne un diagramme en barres horizontal de la catégorie
socio professionnelle en fonction du nombre d'élus de la commune.

Lorsque le dataframe contient plusieurs communes, il retourne Warning :
"Le data.frame contient plusieurs communes. Veuillez filtrer pour une seule commune."
}
\description{
La fonction compte le nombre d'élus par catégorie socio professionnelle
différentes, dans un dataframe qui contient une seule commune et qui valide le schéma
(mêmes noms et nombres de colonnes) du dataframe élus-conseillers-municipaux-cm.
}
