% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction_plot.R
\name{plot.departement}
\alias{plot.departement}
\title{Graphique du nombre d'élus par catégorie socio professionnelle par départements}
\usage{
\method{plot}{departement}(df)
}
\arguments{
\item{df}{Un data.frame de classe "departement" contenant les données des élus.
Doit contenir les colonnes :
\itemize{
\item Code.du.département
\item Libellé.de.la.commune
\item Code.de.la.catégorie.socio.professionnelle
}}
}
\value{
Un objet ggplot2 représentant un diagramme en barres horizontal avec :
\itemize{
\item En ordonnée : les codes des catégories socio-professionnelles
\item En abscisse : le nombre d'élus
\item Un titre incluant le nom du département et le nombre de communes
\item Les barres triées par ordre décroissant du nombre d'élus
\item Les 10 codes professionnels les plus importants en nombre d’élu.e.s du département
}
}
\description{
Cette fonction génère un graphique en barres horizontal représentant la distribution
des élu.e.s municipaux selon leur catégorie socio-professionnelle pour un département donné.
Elle sélectionne automatiquement les 10 catégories les plus représentées en nombre d'élu.e.s.
}
\details{
La fonction réalise les étapes suivantes :
\enumerate{
\item Vérifie la validité du schéma des données
\item Extrait le nom du département et compte le nombre de communes uniques
\item Vérifie qu'il n'y a qu'un seul département dans les données
\item Compte les occurrences de chaque code professionnel
\item Sélectionne les 10 codes les plus fréquents
\item Crée un graphique avec ggplot2 utilisant :
- Une couleur steelblue pour les barres
- Un thème minimal
- Une taille de texte réduite (5) pour les codes sur l'axe Y
}
}
\section{Erreurs}{

La fonction s'arrête avec une erreur si :
\itemize{
\item Le data.frame contient des données de plusieurs départements
\item Le schéma des données n'est pas valide
}
}

\examples{
# Attribution de la classe 'département' au dataframe "mon_dataframe"
mon_dataframe <- creer_departement(mon_dataframe)

# Utilisation de la fonction plot
plot(mon_dataframe)

}
