% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{plot_code_professions}
\alias{plot_code_professions}
\title{Graphique du nombre d'élus par catégorie socio professionnelle}
\usage{
plot_code_professions(df)
}
\arguments{
\item{df}{Un dataframe qui valide le schéma contenant les données
des élus conseillers municipaux.}
}
\value{
La fonction retourne un diagramme en barres horizontal de la catégorie
socio professionnelle en fonction du nombre d'élus.
}
\description{
Cette fonction crée un diagramme en barres horizontal représentant la
distribution des élus selon leur catégorie socio-professionnelle.
Les données doivent provenir d'un dataframe respectant le schéma du
fichier élus-conseillers-municipaux-cm. Les catégories sont triées
par nombre d'élus décroissant.
}
