% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{summary.commune}
\alias{summary.commune}
\title{Résumer des informations des objets de type 'communes'}
\usage{
\method{summary}{commune}(x)
}
\arguments{
\item{x}{Un paramètre x qui sera un objet de type 'commune'.}
}
\value{
La fonction retourne :
Le nom de la commune.
Le nombre d'élu.e.s dans la commune.
La distribution des âges des élu.e.s dans la commune.
Le nom, le prénom et l'âge de l'élu.e le ou la plus âgé.e de la commune.

Lorsque l'objet n'est pas de type commune, la fonction retourne warning :
"L'objet doit être de type 'commune'."
}
\description{
Fonction générique qui sert à résumer les informations des objets de type 'communes'.
Ces objets 'communes' sont des dataframes qui valide le schéma (mêmes noms et nombres de colonnes)
du dataframe élus-conseillers-municipaux-cm. Ces objets doivent être attribué de la classe 'commune',
ce qui permet de reconnaître que ces dataframes doivent être traités par la méthode spécifique
summary.commune lorsque la fonction summary() est appelée sur eux.
}
\examples{
# Affectation de la classe 'commune' au dataframe df_Nantes.
class(df_Nantes) <- c("commune", class(df_Nantes))
# Ne pas exécuter plusieurs fois !

summary(df_Nantes)
summary.commune(df_Nantes)

}
