% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{summary.departement}
\alias{summary.departement}
\title{Résumer des informations des objets de type 'département'}
\usage{
\method{summary}{departement}(x)
}
\arguments{
\item{x}{Un paramètre x qui sera un objet de type 'département'.}
}
\value{
La fonction retourne :
Le nom du département.
Le nombre de commune dans le département.
Le nombre d'élu.e.s dans le département.
La distribution des âges des élu.e.s dans le département.
Le nom, l'âge et la commune de l'élu.e le ou la plus âgé.e du département.
Le nom, l'âge et la commune de l'élu.e le ou la plus jeunes du département.
Le nom de la commune qui à la moyenne d’âge la plus faible, et la distribution des âges des
élu.e.s pour cette commune dans le département.
Le nom de la commune qui à la moyenne d’âge la plus élevé, et la distribution des âges des
élu.e.s pour cette commune dans le département.

Lorsque l'objet n'est pas de type département, la fonction retourne warning :
"L'objet doit être de type 'département'."
}
\description{
Fonction générique qui sert à résumer les informations des objets de type 'département'.
Ces objets 'département' sont des dataframes qui valide le schéma (mêmes noms et nombres de colonnes)
du dataframe élus-conseillers-municipaux-cm. Ces objets doivent être attribué de la classe 'département',
ce qui permet de reconnaître que ces dataframes doivent être traités par la méthode spécifique
summary.departement lorsque la fonction summary() est appelée sur eux.
}
\examples{
# Affectation de la classe 'département' au dataframe df_Gers.
class(df_Gers) <- c("département", class(df_Gers))
# Ne pas exécuter plusieurs fois !

summary(df_Gers)
summary.departement(df_Gers)

}
