% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{trouver_l_elu_le_plus_age}
\alias{trouver_l_elu_le_plus_age}
\title{Trouver l'élu le plus âgé}
\usage{
trouver_l_elu_le_plus_age(df)
}
\arguments{
\item{df}{Un dataframe qui valide le schéma contenant les données
des élus conseillers municipaux.}
}
\value{
Un dataframe d'une ligne contenant le nom, prénom, date de naissance
et âge de l'élu le plus âgé
}
\description{
Cette fonction identifie l'élu le plus âgé dans un dataframe respectant
le schéma des données des élus conseillers municipaux. Elle retourne les
informations de l'élu (nom, prénom, date de naissance et âge calculé).
Le dataframe doit contenir les mêmes colonnes que le fichier
élus-conseillers-municipaux-cm.
}
