% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{validate_schema}
\alias{validate_schema}
\title{Validation du schéma du dataframe élus-conseillers-municipaux-cm}
\usage{
validate_schema(df)
}
\arguments{
\item{df}{Un dataframe contenant les données des élus conseillers municipaux.}
}
\value{
Un message confirmant que le schéma est conforme si la validation réussit.
}
\description{
Cette fonction vérifie que le dataframe fourni respecte le schéma attendu
pour les données des élus conseillers municipaux. Si le schéma ne correspond
pas aux spécifications, la fonction génère une erreur détaillée.
}
\examples{
df <- data.frame(
  "Code.du.département" = character(),
  "Libellé.du.département" = character(),
  "Code.de.la.collectivité.à.statut.particulier" = character(),
  "Libellé.de.la.collectivité.à.statut.particulier" = character(),
  "Code.de.la.commune" = character(),
  "Libellé.de.la.commune" = character(),
  "Nom.de.l.élu" = character(),
  "Prénom.de.l.élu" = character(),
  "Code.sexe" = character(),
  "Date.de.naissance" = character(),
  "Code.de.la.catégorie.socio.professionnelle" = character(),
  "Libellé.de.la.catégorie.socio.professionnelle" = character(),
  "Date.de.début.du.mandat" = character(),
  "Libellé.de.la.fonction" = character(),
  "Date.de.début.de.la.fonction" = character(),
  "Code.nationalité" = character()
)
validate_schema(df)
}
