% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{calcul_distribution_age}
\alias{calcul_distribution_age}
\title{Distribution des âges}
\usage{
calcul_distribution_age(df)
}
\arguments{
\item{df}{Un dataframe qui valide le schéma contenant les données
des élus conseillers municipaux.}
}
\value{
La fonction retourne la distribution des âges dans chaque quantiles
dans le dataframe choisie.
}
\description{
Cette fonction calculera les quantiles 0, 25, 50, 75, 100 de l’âge des élu.e.s.
}
