% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{compter_nombre_d_adjoints}
\alias{compter_nombre_d_adjoints}
\title{Compter le nombre d'adjoints}
\usage{
compter_nombre_d_adjoints(df)
}
\arguments{
\item{df}{Un dataframe qui valide le schéma contenant les données
des élus conseillers municipaux.}
}
\value{
Le nombre d'adjoints.
}
\description{
Cette fonction permet de compter le nombre d'adjoints dans un dataframe
qui valide le schéma (mêmes noms et nombres de colonnes) du dataframe
élus-conseillers-municipaux-cm.
}
