% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Premier script.R
\name{compter_nombre_d_elu}
\alias{compter_nombre_d_elu}
\title{Compter le nombre d'élus}
\usage{
compter_nombre_d_elu(df)
}
\arguments{
\item{df}{= le dataframe élu_conseiller_municipaux}
}
\value{
Le nombre d'élus
}
\description{
Fonction qui permet de compter le nombre d'élus dans le dataframe élu_conseiller_municipaux
}
