% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{compter_nombre_d_elu}
\alias{compter_nombre_d_elu}
\title{Compter le nombre d'élus}
\usage{
compter_nombre_d_elu(df)
}
\arguments{
\item{df}{Un dataframe qui valide le schéma contenant les données
des élus conseillers municipaux.}
}
\value{
Le nombre d'élus.
}
\description{
Cette fonction permet de compter le nombre d'élus dans un dataframe qui
valide le schéma (mêmes noms et nombres de colonnes) du dataframe
élus-conseillers-municipaux-cm.
}
