% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction_plot.R
\name{plot.commune}
\alias{plot.commune}
\title{Graphique du nombre d'élus par catégorie socio professionnelle dans une commune}
\usage{
\method{plot}{commune}(df)
}
\arguments{
\item{df}{Un dataframe qui contient une seule commune et qui valide le schéma
contenant les données des élus conseillers municipaux.}
}
\value{
La fonction retourne un diagramme en barres horizontal de la catégorie
socio professionnelle en fonction du nombre d'élus de la commune.
}
\description{
Cette fonction génère un diagramme en barres horizontal représentant le nombre d'élus
par catégorie socio-professionnelle dans une commune donnée.
Le dataframe fourni doit contenir les données d'une seule commune et respecter le schéma
du fichier \code{élus-conseillers-municipaux-cm} (mêmes noms et nombres de colonnes).
}
\details{
Si le dataframe contient plusieurs communes, la fonction génère une erreur :
\code{"Le data.frame contient plusieurs communes. Veuillez filtrer pour une seule commune."}
}
\examples{
# Attribution de la classe 'commune' au dataframe "mon_dataframe"
mon_dataframe <- creer_commune(mon_dataframe)

# Utilisation de la fonction plot
plot(mon_dataframe)

}
