% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{plot_code_professions}
\alias{plot_code_professions}
\title{Graphique du nombre d'élus par catégorie socio professionnelle}
\usage{
plot_code_professions(df)
}
\arguments{
\item{df}{Un dataframe qui valide le schéma contenant les données
des élus conseillers municipaux.}
}
\value{
La fonction retourne un diagramme en barres horizontal de la catégorie
socio professionnelle en fonction du nombre d'élus.
}
\description{
La fonction compte le nombre d'élus par catégorie socio professionnelle
différentes, dans un dataframe qui valide le schéma (mêmes noms et nombres de colonnes)
du dataframe élus-conseillers-municipaux-cm.
}
