% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction_summary.R
\name{summary.departement}
\alias{summary.departement}
\title{Résumer des informations des objets de type 'département'}
\usage{
\method{summary}{departement}(x)
}
\arguments{
\item{x}{Un objet de classe 'département'.}
}
\value{
La fonction retourne les informations suivantes :
\itemize{
\item Le nom du département.
\item Le nombre de communes dans le département.
\item Le nombre d'élu.e.s dans le département.
\item La distribution des âges des élu.e.s dans le département.
\item Le nom, l'âge et la commune de l'élu.e le ou la plus âgé.e du département.
\item Le nom, l'âge et la commune de l'élu.e le ou la plus jeune du département.
\item La commune ayant la moyenne d’âge la plus basse et la distribution des âges des élu.e.s dans cette commune.
\item La commune ayant la moyenne d’âge la plus élevée et la distribution des âges des élu.e.s dans cette commune.
}
}
\description{
Fonction générique permettant de résumer les informations des objets de type 'département'.
Ces objets sont des dataframes conformes au schéma du fichier \code{élus-conseillers-municipaux-cm}
(mêmes noms et nombres de colonnes) et doivent être attribués de la classe 'département'.
}
\details{
Si l'objet fourni n'est pas de type 'département', la fonction génère un message d'avertissement :
\code{"L'objet doit être de type 'département'."}
}
\examples{
# Attribution de la classe 'département' au dataframe df_Gers
class(df_Gers) <- c("département", class(df_Gers))

summary(df_Gers)
summary.departement(df_Gers)

}
