% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{trouver_l_elu_le_plus_age}
\alias{trouver_l_elu_le_plus_age}
\title{Trouver l'élu le plus âgé}
\usage{
trouver_l_elu_le_plus_age(df)
}
\arguments{
\item{df}{Un dataframe qui valide le schéma contenant les données
des élus conseillers municipaux.}
}
\value{
La fonction retourne l'élus le plus âgé dans le dataframe utilisé.
}
\description{
Cette fonction permet de l'élu le plus âgé dans un dataframe qui valide
le schéma (mêmes noms et nombres de colonnes) du dataframe
élus-conseillers-municipaux-cm.
}
